refactor: [C4R-133] refactor interfaces
