+++ frontend/src/components/RadioButton/RadioButtonConfig.interface.ts
@@ -1,9 +1,8 @@
 import { ChangeEventHandler } from 'react';
 
 export interface RadioButtonConfig {
-  type: 'radio';
   label: string | undefined;
   name: string;
-  value?: string;
+  value: string;
   onChange?: ChangeEventHandler<HTMLInputElement>;
 }
